+++ tortoise/contrib/quart/__init__.py
@@ -77,12 +77,13 @@ def register_tortoise(
     ConfigurationError
         For any configuration error
     """
+    _generate_schemas = generate_schemas
 
     @app.before_serving
     async def init_orm() -> None:  # pylint: disable=W0612
         await Tortoise.init(config=config, config_file=config_file, db_url=db_url, modules=modules)
         logging.info("Tortoise-ORM started, %s, %s", Tortoise._connections, Tortoise.apps)
-        if generate_schemas:
+        if _generate_schemas:
             logging.info("Tortoise-ORM generating schema")
             await Tortoise.generate_schemas()