+++ CHANGELOG.rst
@@ -9,6 +9,13 @@ Changelog
 
 0.16
 ====
+0.16.5
+------
+* Moved ``Tortoise.describe_model(<MODEL>, ...)`` to ``<MODEL>.describe(...)``
+* Deprecated ``Tortoise.describe_model()``
+* Fix for ``generate_schemas`` param being ignored in ``tortoise.contrib.quart.register_tortoise``
+* Fix join query with `source_field` param
+
 0.16.4
 ------
 * More consistent escaping of db columns, fixes using SQL reserved keywords as field names with a function.